requirements.txt
aiohappyeyeballs             2.6.1
aiohttp                      3.12.14
aiohttp-retry                2.9.1
aiosignal                    1.4.0
annotated-types              0.7.0
anyio                        4.9.0
attrs                        25.3.0
cachetools                   5.5.2
certifi                      2025.7.14
cffi                         1.17.1
charset-normalizer           3.4.2
click                        8.2.1
colorama                     0.4.6
dataclasses-json             0.6.7
distro                       1.9.0
fastapi                      0.116.1
filetype                     1.2.0
frozenlist                   1.7.0
google-ai-generativelanguage 0.6.18
google-api-core              2.25.1
google-auth                  2.40.3
google-genai                 1.26.0
googleapis-common-protos     1.70.0
greenlet                     3.2.3
grpcio                       1.73.1
grpcio-status                1.73.1
h11                          0.16.0
httpcore                     1.0.9
httpx                        0.28.1
httpx-sse                    0.4.1
idna                         3.10
iniconfig                    2.1.0
jiter                        0.10.0
jsonpatch                    1.33
jsonpointer                  3.0.0
langchain                    0.3.26
langchain-community          0.3.27
langchain-core               0.3.69
langchain-google-genai       2.1.8
langchain-openai             0.3.28
langchain-pinecone           0.2.9
langchain-tests              0.3.20
langchain-text-splitters     0.3.8
langgraph                    0.5.3
langgraph-checkpoint         2.1.1
langgraph-prebuilt           0.5.2
langgraph-sdk                0.1.73
langsmith                    0.4.6
markdown-it-py               3.0.0
marshmallow                  3.26.1
mdurl                        0.1.2
multidict                    6.6.3
mypy_extensions              1.1.0
numpy                        2.3.1
openai                       1.97.0
orjson                       3.11.0
ormsgpack                    1.10.0
packaging                    24.2
pillow                       11.3.0
pinecone                     7.3.0
pinecone-client              6.0.0
pinecone-plugin-assistant    1.7.0
pinecone-plugin-interface    0.0.7
pip                          25.1.1
pluggy                       1.6.0
propcache                    0.3.2
proto-plus                   1.26.1
protobuf                     6.31.1
py-cpuinfo                   9.0.0
pyasn1                       0.6.1
pyasn1_modules               0.4.2
pycparser                    2.22
pydantic                     2.11.7
pydantic_core                2.33.2
pydantic-settings            2.10.1
Pygments                     2.19.2
pytesseract                  0.3.13
pytest                       8.4.1
pytest-asyncio               0.26.0
pytest-benchmark             5.1.0
pytest-codspeed              4.0.0
pytest-recording             0.13.4
pytest-socket                0.7.0
python-dateutil              2.9.0.post0
python-dotenv                1.1.1
python-multipart             0.0.20
PyYAML                       6.0.2
regex                        2024.11.6
requests                     2.32.4
requests-toolbelt            1.0.0
rich                         14.0.0
rsa                          4.9.1
six                          1.17.0
sniffio                      1.3.1
SQLAlchemy                   2.0.41
starlette                    0.47.1
syrupy                       4.9.1
tenacity                     8.5.0
tiktoken                     0.9.0
tqdm                         4.67.1
typing_extensions            4.14.1
typing-inspect               0.9.0
typing-inspection            0.4.1
urllib3                      2.5.0
uvicorn                      0.35.0
vcrpy                        7.0.0
websockets                   15.0.1
wrapt                        1.17.2
xxhash                       3.5.0
yarl                         1.20.1
zstandard                    0.23.0
